updated dots packages to latest

+++ Assets/Scripts/Action Framework/Frame Player/Runtime Data/OnPlayUpdate.cs
@@ -4,5 +4,5 @@ public struct OnPlayUpdate : IComponentData
 {
     public bool loop;
     public int currentFrame;
-    public double normlizedTime;
+    public float normlizedTime;
 }
